update badge to render customised chip component
