+++ src/components/Badge/index.cy.js
