+++ src/pages/trees/[treeid].js
@@ -10,8 +10,8 @@ import Typography from '@mui/material/Typography';
 import log from 'loglevel';
 import moment from 'moment';
 import { useRouter } from 'next/router';
-import { useEffect } from 'react';
-import Badges from 'components/Badges';
+import { useEffect, useMemo } from 'react';
+import Badge from 'components/Badge';
 import HeadTag from 'components/HeadTag';
 import InformationCard1 from 'components/InformationCard1';
 import LikeButton from 'components/LikeButton';
@@ -179,17 +179,25 @@ export default function Tree({
 
   log.warn(planter, 'planter');
 
-  const badgesContent = [
-    {
-      color: tree?.approved ? 'primary' : 'greyLight',
-      icon: tree?.approved ? <CheckIcon /> : null,
-      badgeName: tree?.approved ? 'Waiting for verification' : 'Verified',
-    },
-    {
-      color: 'secondary',
-      badgeName: tree?.tokenId ? 'Token not issued' : 'Token issued',
-    },
-  ];
+  // storing under variable with useMemo wrapped
+  // to reuse the same component for mobile and desktop and
+  // avoid re-rendering of badge components
+  const BadgeSection = useMemo(
+    () => (
+      <>
+        <Badge
+          color={tree?.approved ? 'primary' : 'greyLight'}
+          icon={tree?.approved ? <CheckIcon /> : null}
+          badgeName={tree?.approved ? 'Waiting for verification' : 'Verified'}
+        />
+        <Badge
+          color="secondary"
+          badgeName={tree?.token_id ? 'Token not issued' : 'Token issued'}
+        />
+      </>
+    ),
+    [tree?.approved, tree?.token_id],
+  );
 
   return (
     <>
@@ -297,7 +305,7 @@ export default function Tree({
                   mt: 2,
                 }}
               >
-                <Badges content={badgesContent} />
+                {BadgeSection}
               </Box>
             </Box>
           </Portal>
@@ -541,7 +549,7 @@ export default function Tree({
                     mt: 2,
                   }}
                 >
-                  <Badges content={badgesContent} />
+                  {BadgeSection}
                 </Box>
               </Box>
             )}
