+++ src/pages/v2/captures/[captureid].js
@@ -1,5 +1,6 @@
 /* eslint-disable @next/next/no-img-element */
 import AccountBalanceWalletIcon from '@mui/icons-material/AccountBalanceWallet';
+import CheckIcon from '@mui/icons-material/Check';
 import HubIcon from '@mui/icons-material/Hub';
 import VerifiedIcon from '@mui/icons-material/Verified';
 import { useTheme, Avatar, Divider } from '@mui/material';
@@ -9,8 +10,8 @@ import Typography from '@mui/material/Typography';
 import log from 'loglevel';
 import moment from 'moment';
 import { useRouter } from 'next/router';
-import { useEffect } from 'react';
-import Badges from 'components/Badges';
+import { useEffect, useMemo } from 'react';
+import Badge from 'components/Badge';
 import HeadTag from 'components/HeadTag';
 import ImpactSection from 'components/ImpactSection';
 import InformationCard1 from 'components/InformationCard1';
@@ -34,12 +35,7 @@ import TokenIcon from 'images/icons/token.svg';
 import imagePlaceholder from 'images/image-placeholder.png';
 import SearchIcon from 'images/search.svg';
 import { useMapContext } from 'mapContext';
-import {
-  getStakeHolderById,
-  getCapturesById,
-  getGrowerById,
-  getCountryByLatLon,
-} from 'models/api';
+import { getStakeHolderById, getCapturesById, getGrowerById, getCountryByLatLon } from 'models/api';
 import * as pathResolver from 'models/pathResolver';
 import * as utils from 'models/utils';
 
@@ -186,6 +182,31 @@ export default function Capture({
 
   log.warn(grower, 'grower');
 
+  // storing under variable with useMemo wrapped
+  // to reuse the same component for mobile and desktop and
+  // avoid re-rendering of badge components
+  const BadgeSection = useMemo(
+    () => (
+      <>
+        <Badge
+          color={tree?.approved ? 'primary' : 'greyLight'}
+          icon={tree?.approved ? <CheckIcon /> : null}
+          badgeName={tree?.approved ? 'Waiting for verification' : 'Verified'}
+        />
+        <Badge
+          color="secondary"
+          badgeName={tree?.token_id ? 'Token not issued' : 'Token issued'}
+        />
+        <Badge
+          color={tree.id ? 'primary' : 'greyLight'}
+          badgeName={tree.id ? 'Tree matched' : 'Waiting for tree match'}
+          onClick={tree.id ? () => router.push(`/trees/${tree.id}`) : null}
+        />
+      </>
+    ),
+    [tree?.approved, tree?.token_id, tree?.id],
+  );
+
   return (
     <>
       <HeadTag title={`Tree #${tree.id}`} />
@@ -291,7 +312,7 @@ export default function Capture({
                   mt: 2,
                 }}
               >
-                <Badges tokenId={tree.token_id} verified={tree.verified} />
+                {BadgeSection}
               </Box>
             </Box>
           </Portal>
@@ -526,7 +547,7 @@ export default function Capture({
                   mt: 2,
                 }}
               >
-                <Badges tokenId={tree.token_id} verified={tree.verified} />
+                {BadgeSection}
               </Box>
             </Box>
           )}
@@ -692,7 +713,7 @@ export default function Capture({
             mt: [10, 20],
           }}
         />
-       
+
         <Box height={20} />
         {nextExtraIsEmbed && (
           <Portal
